move get context from brain
