--- backend/repository/brain/__init__.py
@@ -1,1 +1,0 @@
-from .get_question_context_from_brain import get_question_context_from_brain
